including user authorization in backend with new ref and userid to backend model and posts; passing the userId to fron-end; using the userId to hide edit and delete button for non creators.

+++ backend/routes/posts.js
@@ -40,8 +40,8 @@ router.post('', checkAuth, multer({ storage: storage }).single("image"), (req, r
   const post = new Post({
     title: req.body.title,
     content: req.body.content,
-    imagePath: url + "/images/" + req.file.filename
-
+    imagePath: url + "/images/" + req.file.filename,
+    creator: req.userData.userId
   })
   // console.log(post);
   post.save().then(createdPost => {
@@ -107,18 +107,30 @@ router.put('/:id', checkAuth, multer({ storage: storage }).single("image"), (req
     _id: req.body.id,
     title: req.body.title,
     content: req.body.content,
-    imagePath: imagePath
+    imagePath: imagePath,
+    creator: req.userData.userId
+
   });
-  Post.updateOne({ _id: req.params.id }, post).then((result) => {
+  Post.updateOne({ _id: req.params.id, creator: req.userData.userId }, post).then((result) => {
+    if (result.nModified > 0) {
+      res.status(200).json({ message: "Update successful!" });
+    } else {
+      res.status(401).json({ message: "Not authorized!" });
+    }
     console.log(result);
-    res.status(200).json({ message: "Post updated successfully" })
+    // res.status(200).json({ message: "Post updated successfully" })
   })
 });
 router.delete('/:id', checkAuth, (req, res, next) => {
 
-  Post.deleteOne({ _id: req.params.id }).then(result => {
+  Post.deleteOne({ _id: req.params.id, creator: req.userData.userId }).then(result => {
     console.log(result);
-    res.status(200).json({ message: "Posts deleted successfully!" });
+    if (result.n > 0) {
+      res.status(200).json({ message: "Deletion successful!" });
+    } else {
+      res.status(401).json({ message: "Not authorized!" });
+    }
+    // res.status(200).json({ message: "Posts deleted successfully!" });
   })
 
 })
